The type of zone wasn't show to users with access level 1.
If a user with access level 1 did have access to a slave zone the
user did not see the IP of the master nameserver of that zone. Now 
the user will the IP address (readonly and only if one is set).

Bugfix. If no master IP for a slave zone is given, a warning is
shown regardless of the userlevel.

Bugfix. Both the "add record" and "edit record" buttons in the "edit
zone" screen is no longer available for users with level 1 for
domains of type "slave". Bug report by Antonio Prado.

Some PHP and HTML cleanup (removing of empty tags and unnecessary 
repeatings of calls to a single function).