Implement the logging of "delete single domain"
When the user deletes a domain in the user interface, the deletion of
the domain as well as the deletion of all records in that domain is
logged.

Zones (poweradmin concept) were renamed to Domains (powerdns concept).

A new column `log_domains`.`domain_name` was added to log which
domain was removed.

RecordLog and DomainLog now support injection of the database
variable with (Domain|Record)Log::with_db($db).